Moved new post javascript to its own file
