+++ app/templates/new_meeting_notes.html
@@ -117,4 +117,5 @@
 
 {% block custom_scripts %}
     <script type="text/javascript" src="{{ url_for('static', filename='js/plugins/bootstrap-multiselect.js') }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='js/new_post.js') }}"></script>
 {% endblock %}
